feat: Push the 'Input' rendering code into Input.tsx
